milvus_server/requirements.txt
fastapi
uvicorn
requests
numpy
etcd3
minio
pulsar-client
